requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
decorator==4.4.2
idna==3.2
imageio==2.9.0
imageio-ffmpeg==0.4.4
moviepy==1.0.3
numpy==1.19.5
Pillow==8.3.1
proglog==0.1.9
python-dotenv==0.19.0
pytube==11.0.0
requests==2.26.0
tqdm==4.62.1
urllib3==1.26.6
youtube-transcript-api==0.4.1
